requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
accelerate==1.13.0
    # via -r requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via fsspec
aiosignal==1.4.0
    # via aiohttp
altair==6.0.0
    # via streamlit
annotated-doc==0.0.4
    # via typer
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via httpx
async-timeout==5.0.1
    # via aiohttp
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
bitsandbytes==0.49.2
    # via -r requirements.in
blinker==1.9.0
    # via streamlit
cachetools==7.0.3
    # via streamlit
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.5
    # via requests
click==8.3.1
    # via
    #   streamlit
    #   typer
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
datasets==4.6.1
    # via -r requirements.in
dill==0.4.0
    # via
    #   datasets
    #   multiprocess
exceptiongroup==1.3.1
    # via anyio
filelock==3.25.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
fonttools==4.61.1
    # via matplotlib
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2026.2.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
h11==0.16.0
    # via httpcore
hf-xet==1.3.2
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   datasets
    #   huggingface-hub
huggingface-hub==1.6.0
    # via
    #   accelerate
    #   datasets
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.5.3
    # via scikit-learn
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
kiwisolver==1.4.9
    # via matplotlib
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.8
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.18
    # via datasets
narwhals==2.17.0
    # via altair
networkx==3.4.2
    # via torch
numpy==2.2.6
    # via
    #   -r requirements.in
    #   accelerate
    #   bitsandbytes
    #   contourpy
    #   datasets
    #   matplotlib
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   streamlit
    #   transformers
packaging==26.0
    # via
    #   accelerate
    #   altair
    #   bitsandbytes
    #   datasets
    #   huggingface-hub
    #   matplotlib
    #   streamlit
    #   transformers
pandas==2.3.3
    # via
    #   -r requirements.in
    #   datasets
    #   streamlit
pillow==12.1.1
    # via
    #   matplotlib
    #   streamlit
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.5
    # via streamlit
psutil==7.2.2
    # via accelerate
pyarrow==23.0.1
    # via
    #   datasets
    #   streamlit
pydantic==2.12.5
    # via -r requirements.in
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via rich
pyparsing==3.3.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2026.1.post1
    # via pandas
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   transformers
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2026.2.28
    # via transformers
requests==2.32.5
    # via
    #   datasets
    #   streamlit
rich==14.3.3
    # via typer
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
safetensors==0.7.0
    # via
    #   accelerate
    #   transformers
scikit-learn==1.7.2
    # via -r requirements.in
scipy==1.15.3
    # via scikit-learn
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.55.0
    # via -r requirements.in
sympy==1.14.0
    # via torch
tenacity==9.1.4
    # via streamlit
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.2
    # via transformers
toml==0.10.2
    # via streamlit
torch==2.10.0
    # via
    #   -r requirements.in
    #   accelerate
    #   bitsandbytes
tornado==6.5.4
    # via streamlit
tqdm==4.67.3
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   transformers
transformers==5.3.0
    # via -r requirements.in
typer==0.24.1
    # via
    #   huggingface-hub
    #   transformers
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   altair
    #   anyio
    #   exceptiongroup
    #   huggingface-hub
    #   multidict
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
xxhash==3.6.0
    # via datasets
yarl==1.23.0
    # via aiohttp
